fix!: wrap all data in AFEnumerationValue if possible
- Moved AFEnumerationValue from Asset to Data to prevent cyclical import
